Fixed all slow downs for Ring of Fire!
I added the buffer which allowed me to find out that I set the part of the trail with half the lifetime as the collision incorrectly. There is now no slow down even without the buffer :). I know starting attempting to make the size of the trail attack circle to scale with the distance between point A and B.