Mace Adjusts Begin
Added in some of the new sprites and worked to adjuest their collision bozes as well as the maces. Figured out the length for rope, need to do a bit more digging into how the rop works to be able to apply a new color or a material/texture to the rope instead of it just being a glowing pink line of light. Though that might also work as an aesthetic.